+++ assets/javascripts/kitten/components/cards/crowdfunding-card/stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { withKnobs, text, boolean, number } from '@storybook/addon-knobs'
+import { withKnobs, text, boolean, number, select } from '@storybook/addon-knobs'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
 import { Grid, GridCol } from '../../grid/grid'
@@ -9,11 +9,12 @@ import { CrowdfundingCardWidget as CrowdfundingCardWidgetComponent } from './wid
 import { KissKissBankBankIcon } from '../../../components/icons/kisskissbankbank-icon'
 import { Text } from '../../../components/typography/text'
 import { pxToRem } from '../../../helpers/utils/typography'
+import { STATE_CHOICES } from './stories/project-state'
 
 const StoryContainer = ({ children }) => (
   <Container>
     <Grid>
-      <GridCol col-m="6">
+      <GridCol col-m="4">
         <Marger top="5" bottom="5">
           {children}
         </Marger>
@@ -101,3 +102,36 @@ export const CrowdfundingCardWidget = () => (
     />
   </StoryContainer>
 )
+
+const options = Object.keys(STATE_CHOICES)
+
+export const LendopolisCrowdfundingCard = () => {
+  const widgetState = select('widget State', options, options[0])
+
+  return (
+    <StoryContainer>
+      <CrowdfundingCardComponent
+        href={text('Link', '#')}
+        ownerTitle={text('Owner title', 'Owner title')}
+        ownerDescription={text('Owner description', 'Owner description')}
+        loading={boolean('Loading', false)}
+        stretch={boolean('Stretch', false)}
+        cardTitle={text(
+          'Title',
+          'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do',
+        )}
+        cardSubTitle={text(
+          'SubTitle',
+          'Sed ut perspiciatis unde omnis iste natus error Lorem',
+        )}
+        info1={text('Info1', 'Lorem ipsum…')}
+        info2={text('Info2', 'Lorem ipsum…')}
+        info3={text('Info3', 'Lorem ipsum…')}
+        progress={number('Progress', 42)}
+        widgetState={STATE_CHOICES[widgetState]}
+        additionalInfo={"Collecte réservée aux départements : 92, 34, 12, 82"}
+        title="Aller sur la collecte …"
+      />
+    </StoryContainer>
+  )
+}