+++ assets/javascripts/kitten/components/icons/flash-circle-icon.js
@@ -1,17 +1,17 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import COLORS from '../../../constants/colors-config'
-import { computeFromRatio } from '../../../helpers/utils/ratio'
+import COLORS from '../../constants/colors-config'
+import { computeFromRatio } from '../../helpers/utils/ratio'
 
 const DEFAULT_WIDTH = 20
 const DEFAULT_HEIGHT = 20
 
-export const CheckedCircleIcon = ({
+export const FlashCircleIcon = ({
   circleColor,
-  checkedColor,
-  title,
+  flashColor,
   width,
   height,
+  title,
   ...props
 }) => {
   const computed = computeFromRatio({
@@ -32,26 +32,23 @@ export const CheckedCircleIcon = ({
       {title && <title>{title}</title>}
       <circle fill={circleColor} cx="10" cy="10" r="10" />
       <path
-        fill={checkedColor}
-        d="M8.232 13.89l.707.706 5.656-5.657-1.414-1.415-4.243 4.243-2.122-2.122-1.414 1.415 2.828 2.83z"
+        d="m6 10 6-6-1.25 6h3.25l-6 6 1.25-6z"
+        fill={flashColor}
+        fillRule="evenodd"
       />
     </svg>
   )
 }
 
-CheckedCircleIcon.prototype = {
-  color: PropTypes.string,
+FlashCircleIcon.prototype = {
+  circleColor: PropTypes.string,
+  flashColor: PropTypes.string,
   title: PropTypes.string,
   width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   height: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
 }
 
-CheckedCircleIcon.defaultProps = {
+FlashCircleIcon.defaultProps = {
   circleColor: COLORS.background1,
-  checkedColor: COLORS.font1,
-  title: 'Checked circle',
-}
-
-export const checkedCircleIconAsString = ({ circleColor, checkedColor }) => {
-  return `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><circle fill="${circleColor}" cx="10" cy="10" r="10" /><path fill="${checkedColor}" d="M8.232 13.89l.707.706 5.656-5.657-1.414-1.415-4.243 4.243-2.122-2.122-1.414 1.415 2.828 2.83z"/></svg>`
+  flashColor: COLORS.font1,
 }